+++ src/characters/services/characters.service.ts
@@ -21,6 +21,7 @@ export class CharactersService {
           params,
         },
       );
+
       return data;
     } catch (error) /* istanbul ignore next */ {
       throw new HttpException(error.message, error.response.status);